move openInBrowser back out and regen cycles file
