+++ src/design-system/components/TextLink/TextLink.tsx
@@ -1,7 +1,7 @@
 import React, { useCallback, type ReactNode } from 'react';
 import { Text as NativeText, type TextStyle } from 'react-native';
 
-import { openInBrowser } from '@/features/dapp-browser/utils/openInBrowser';
+import { openInBrowser } from '@/utils/openInBrowser';
 
 const style: TextStyle = {
   textDecorationLine: 'underline',
