+++ src/screens/SettingsSheet/components/support/SupportSection.tsx
@@ -2,13 +2,13 @@ import React, { useMemo } from 'react';
 
 import ButtonPressAnimation from '@/components/animations/ButtonPressAnimation';
 import { Box, Stack, Text } from '@/design-system';
-import { openInBrowser } from '@/features/dapp-browser/utils/openInBrowser';
 import { opacity } from '@/framework/ui/utils/opacity';
 import { useSendFeedback } from '@/hooks/useSendFeedback';
 import * as i18n from '@/languages';
 import { useNavigation } from '@/navigation/Navigation';
 import Routes from '@/navigation/routesNames';
 import { useTheme } from '@/theme/ThemeContext';
+import { openInBrowser } from '@/utils/openInBrowser';
 
 import { SettingsExternalURLs } from '../../constants';
 import Menu from '../Menu';
